python-api/requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
AWSIotPythonSDK==1.5.4
# bcrypt==4.2.0 # NEED TO HAVE BUILD TOOLS
boto3==1.35.17
botocore==1.35.17
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decouple==0.0.7
deprecation==2.1.0
et-xmlfile==1.1.0
fastapi==0.115.0
frozenlist==1.4.1
gotrue==2.9.0
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
imagetk==0.1.2 # NEED TO HAVE BUILD TOOLS
jmespath==1.0.1
keyboard==0.13.5
multidict==6.1.0
mypy==1.11.2
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==9.0.0
mysqlclient==2.2.4
numpy==2.1.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pillow==10.4.0
postgrest==0.16.11
py-bcrypt==0.4 # NEED TO HAVE BUILD TOOLS
pydantic==2.9.2
pydantic_core==2.23.4
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==308
PyYAML==6.0.2
realtime==2.0.5
requests==2.32.3
s3transfer==0.10.2
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
tenacity==9.0.0
tk==0.1.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.31.0
watchfiles==0.24.0
websockets==13.1
yarl==1.13.1
